chore: update Makefile target to be silent and another one to delete bonus binary, update headers
