+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: caide-so <caide-so@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/22 19:11:14 by caide-so          #+#    #+#             */
-/*   Updated: 2025/06/16 00:27:24 by caide-so         ###   ########.fr       */
+/*   Updated: 2025/06/17 02:29:31 by caide-so         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,9 @@
 
 // color macros
 # include "colors.h"
+
+// structs header
+# include "structs.h"
 
 // libft
 # include "../libft/include/libft.h"
@@ -25,7 +28,8 @@
 // to use malloc, free, exit
 # include <stdlib.h>
 
-// to use write, usleep, fork
+// to use write, usleep, fork, getcwd, chdir, unlink, execve, dup, dup2, 
+// pipe, access, read, close, isatty, ttyname, ttyslot, ioctl
 # include <unistd.h>
 
 // for O_* constants
@@ -34,13 +38,11 @@
 // to use pid_t type
 # include <sys/types.h>
 
-// to use waitpid
+// to use waitpid, wait, wait3, wait4
 # include <sys/wait.h>
 
-// to use readline
+// to use readline functions
 # include <readline/readline.h>
-
-// to use add_history
 # include <readline/history.h>
 
 // to use errno and ENOENT
@@ -52,153 +54,32 @@
 // to use signal functions
 # include <signal.h>
 
-// to use stat
+// to use stat, lstat, fstat
 # include <sys/stat.h>
 
 // to use tcsetattr, tcgetattr
 # include <termios.h>
 
-typedef struct sigaction	t_sig;
+// to use strerror
+# include <string.h>
 
-typedef enum e_token_type
-{
-	WORD,
-	PIPE,
-	REDIR_IN,
-	REDIR_OUT,
-	REDIR_APPEND,
-	HEREDOC,
-	LOGICAL_AND,
-	LOGICAL_OR,
-	PAREN_OPEN,
-	PAREN_CLOSE,
-}	t_token_type;
+// to use perror (also in stdio.h)
+// to use getenv
+# include <stdlib.h>
 
-typedef enum e_quote
-{
-	NO_QUOTE,
-	SINGLE_QUOTE,
-	DOUBLE_QUOTE
-}	t_quote;
+// to use termcap functions
+# include <term.h>
 
-typedef enum s_node_type
-{
-	NODE_COMMAND,
-	NODE_PIPE,
-	NODE_AND,
-	NODE_OR,
-	NODE_SUBSHELL
-}	t_node_type;
-
-typedef enum s_redir_type
-{
-	R_IN,
-	R_OUT,
-	R_APPEND,
-	R_HEREDOC
-}	t_redir_type;
-
-typedef struct s_redir
-{
-	t_redir_type	type;
-	char			*filename;
-	int				heredoc_fd;
-	struct s_redir	*next;
-}	t_redir;
-
-typedef struct s_command
-{
-	char			**args;
-	int				arg_count;
-	t_redir			*redirs;
-	int				heredoc_fd;
-	int				is_builtin;
-}	t_command;
-
-typedef struct s_token
-{
-	t_token_type	type;
-	char			*value;
-	struct s_token	*next;
-}	t_token;
-
-typedef struct s_token_list
-{
-	t_token	*head;
-	t_token	*tail;
-	int		size;
-}	t_token_list;
-
-typedef struct s_env
-{
-	char			*key;
-	char			*value;
-	struct s_env	*next;
-}	t_env;
-
-typedef struct s_ast
-{
-	t_node_type		type;
-	struct s_ast	*left;
-	struct s_ast	*right;
-	t_command		*cmd;
-}	t_ast;
-
-typedef struct s_args
-{
-	char			*arg;
-	struct s_args	*next;
-}	t_args;
-
-typedef struct s_expand
-{
-	t_args	**head;
-	t_quote	state;
-	t_env	*env;
-	char	*prefix;
-	char	*cur;
-	int		i;
-}	t_exp;
-
-typedef struct s_wild
-{
-	char	**parts;
-	int		have_start;
-	int		have_end;
-	int		full;
-}	t_wild;
-
-typedef struct s_shell
-{
-	t_env			*env;
-	t_token_list	*tokens;
-	t_ast			*ast;
-	int				last_status;
-}	t_shell;
-
-typedef struct s_quote_ctx
-{
-	char	*str;
-	char	*result;
-	int		i;
-	int		j;
-	t_quote	state;
-}	t_quote_ctx;
-
-extern volatile int			g_signal_status;
+// to use ioctl
+# include <sys/ioctl.h>
 
 // tokenizer
 t_token_list	*tokenizer(char *input);
 int				handle_operators(char *input, int *i, t_token_list *tokens);
 int				in(const char *s, char c);
-
-// token
 t_token			*new_token(t_token_type type, char *value);
-
-// token list
 void			token_list_init(t_token_list *list);
 void			token_list_append(t_token_list *list, t_token *token);
-void			token_list_free(t_token_list *list);
 
 // env
 t_env			*init_env(char **envp);
@@ -207,21 +88,27 @@ char			*get_env_path(t_env *env);
 t_env			*env_new(const char *key, const char *value);
 int				env_append(t_env **head, t_env **tail, t_env *node);
 void			env_update(t_env **env, char *key, char *value);
+char			**env_list_to_array(t_env *env);
+void			free_string_array(char **arr);
 void			env_export_only(t_env **env_head, char *name);
 
 // clean
 void			clean_all(t_token_list *tokens, t_ast *node, t_env **env);
-void			env_free_all(t_env **head);
-void			exit_perror(const char *msg);
 void			report_unexpected_quotes(const char token_value);
 int				report_unexpected(const char *token_value);
 
-// clean_ast
+// clean token list
+void			token_list_free(t_token_list *list);
+
+// clean env
+void			env_free_all(t_env **head);
+
+// clean ast
 void			free_args(t_command *command);
 void			command_free(t_command *command);
 void			ast_free(t_ast *root);
 
-// clean_expansion
+// clean expansion
 void			free_args_temp(char **args);
 void			free_args_list(t_args *args);
 void			free_array(void **ptr);
@@ -230,47 +117,48 @@ void			free_array(void **ptr);
 void			expander(char ***args, t_env *env, int *size_args, t_shell *sh);
 char			**list_to_args(t_args *args, int *size_args);
 
-// expand_env
+// expand env
 t_args			*expand_env(char **args, t_env *env, t_shell *sh);
 char			*start_prefix(char *input, int *i, t_quote *state);
 void			expand_variable(t_exp *exp, char *key);
 void			expand_status(t_exp *exp, int status);
 void			handle_no_quotes(t_exp *exp);
-
-//expand_token
-t_args			*expand_token(char *input, t_env *env, t_shell *sh);
-void			init_expander(t_exp *exp, char *input, t_env *env);
-void			handle_next_token(t_exp *exp, char *input, t_shell *sh);
-
-// expand_env_utils
 char			*extract_key(char *input, int *i);
 void			update_state_quote(t_quote *state, char c);
 char			*find_env_value(char *key, t_env *env);
 char			*ft_strjoin_free(char *s1, char *s2);
 
-// expansion_utils
+// expand token
+t_args			*expand_token(char *input, t_env *env, t_shell *sh);
+void			init_expander(t_exp *exp, char *input, t_env *env);
+void			handle_next_token(t_exp *exp, char *input, t_shell *sh);
+
+// expansion utils
 void			add_token(t_args **head, char *value);
 void			append_list(t_args **head, t_args *list);
 int				count_array(void **arr);
 void			add_token_free(t_args **head, char *value);
 
-// expand_wild
+// expand wild
 t_args			*expand_wild(t_args *envs);
 t_wild			parse_pattern(char *arg);
 t_args			*wild_matches(t_wild *wild);
 int				is_wildcard(char *arg);
 int				only_asterisk(char *arg);
-
-// expand_wild_utils
 int				match_middle(char *pos, t_wild *wild, int i, int limit);
 int				match_end(const char *filename, const char *end);
 int				match_pattern(const char *filename, t_wild *wild);
 int				handle_opening_quote(char *input, int *i, t_quote *state);
 int				update_state_quote_i(char *input, int *i, t_quote *state);
-
-// expand_wild_split
 char			**split_wildcard(char *input);
 void			add_token_to_array(char ***array, char *token);
+
+// sort wildcard
+void			swap_args(t_args *a, t_args *b);
+int				ft_tolower(int c);
+int				ft_strcasecmp(const char *s1, const char *s2);
+void			sort_args_list(t_args *list);
+void			sort_append(t_args **head, t_args *list);
 
 // expand redirs
 void			expand_redirs(t_redir *redirs, t_shell *sh);
@@ -281,15 +169,9 @@ void			handle_no_match(t_shell *sh, t_redir *redir);
 void			handle_ambiguous(t_shell *sh, t_redir *redir);
 void			handle_single_match(t_redir *redir, t_args *expanded_list);
 
-// handle_quotes
+// utils
+void			exit_perror(const char *msg);
 char			*rem_quotes(char *str, int free_str);
-
-// sort_wildcard
-void			swap_args(t_args *a, t_args *b);
-int				ft_tolower(int c);
-int				ft_strcasecmp(const char *s1, const char *s2);
-void			sort_args_list(t_args *list);
-void			sort_append(t_args **head, t_args *list);
 
 // ast
 t_ast			*parse_command(t_token **token, t_shell *shell);
@@ -297,6 +179,10 @@ t_ast			*parse_logical(t_token **token, t_shell *shell);
 t_ast			*parse_pipe(t_token **token, t_shell *shell);
 t_ast			*parse_simple_command(t_token **token, t_shell *shell);
 t_ast			*parse_subshell(t_token **token, t_shell *shell);
+int				is_pipe_or_logical(t_token *token);
+int				is_redirect(t_token *token);
+int				get_size_args(t_token **token);
+int				have_quotes(char *raw);
 
 // commands
 t_command		*make_command(t_token **token, t_shell *shell);
@@ -305,21 +191,6 @@ void			parse_redirect(t_token **token, t_command **cmd, t_shell *sh);
 void			parse_heredoc(t_token **token, t_command **cmd, t_shell *shell);
 char			**get_args(t_token **token, int size_args);
 void			append_redir(t_redir **head, t_redir_type type, char *filename);
-
-// ast_utils
-int				is_pipe_or_logical(t_token *token);
-int				is_redirect(t_token *token);
-int				get_size_args(t_token **token);
-int				have_quotes(char *raw);
-
-// debug functions
-void			print_env(t_env *env);
-void			print_envp(char **envp);
-void			print_tokens(t_token_list *tokens);
-void			print_token(char *str_type, t_token *token);
-void			test_expander(t_env *env);
-void			test_commands_from_tokens(t_token_list *tokens);
-void			print_ast(t_ast *node, int level, t_env *env, t_shell *sh);
 
 // syntax analysis
 int				syntax_analysis(t_token_list *tokens);
@@ -348,6 +219,8 @@ void			handle_dir_case(char *cmd, t_shell *shell);
 void			handle_not_found(char *cmd, t_shell *shell);
 void			handle_permission_denied(char *cmd, t_shell *shell);
 int				exec_with_redirs(char **args, t_command *cmd, t_shell *shell);
+void			handle_left_child(int *fd, t_shell *shell, t_ast *node);
+void			handle_right_child(int *fd, t_shell *shell, t_ast *node);
 
 // redir
 int				apply_redirections(t_command *cmd, t_shell *sh);
@@ -366,17 +239,9 @@ char			*ft_strjoin_char_free(char *str, char c);
 int				save_fds(int *save_stdin, int *save_stdout);
 int				restore_fds(int save_stdin, int save_stdout);
 
-// env manipulation
-char			**env_list_to_array(t_env *env);
-void			free_string_array(char **arr);
-
 // last status
 void			set_last_status(t_shell *shell, int status);
 int				get_last_status(t_shell *shell);
-
-// executor pipe
-void			handle_left_child(int *fd, t_shell *shell, t_ast *node);
-void			handle_right_child(int *fd, t_shell *shell, t_ast *node);
 
 // builtin
 void			builtin_exit(t_token_list *tokens, t_shell *shell);
@@ -389,13 +254,13 @@ int				builtin_export(char **args, t_env **env);
 void			print_exported(t_env *env);
 int				builtin_unset(char **args, t_env **env);
 
-// setup_signals_prompt
+// signals prompt
 void			setup_signals_prompt(void);
 t_sig			setup_sigint_prompt(void);
 void			handle_sigint_prompt(int sig);
 void			update_sh_last_status(t_shell *sh, int new_value);
 
-// setup_signals_heredoc
+// signals heredoc
 // void			handle_sigeof_heredoc(char *delim);
 void			handle_sigeof_heredoc(char *delim);
 void			handle_sigint_heredoc(int sig);
@@ -403,7 +268,7 @@ void			setup_signals_heredoc(void);
 void			disable_echoctl(void);
 void			enable_echoctl(void);
 
-// setup_signals_exec
+// signals exec
 void			setup_signals_exec(void);
 t_sig			setup_sigint_exec(void);
 t_sig			setup_sigquit_exec(void);
@@ -420,5 +285,14 @@ void			build_cwd(char *dst, size_t *i, char *cwd);
 void			append_str(char *dst, size_t *i, const char *src);
 void			append_char(char *dst, size_t *i, char c);
 char			*replace_home(char *cwd, char *home);
+
+// debug functions
+void			print_env(t_env *env);
+void			print_envp(char **envp);
+void			print_tokens(t_token_list *tokens);
+void			print_token(char *str_type, t_token *token);
+void			test_expander(t_env *env);
+void			test_commands_from_tokens(t_token_list *tokens);
+void			print_ast(t_ast *node, int level, t_env *env, t_shell *sh);
 
 #endif
