+++ include/structs.h
@@ -1,0 +1,145 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   structs.h                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: caide-so <caide-so@student.42sp.org.br>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/06/17 01:51:16 by caide-so          #+#    #+#             */
+/*   Updated: 2025/06/17 01:54:19 by caide-so         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef STRUCTS_H
+# define STRUCTS_H
+
+typedef struct sigaction	t_sig;
+
+typedef enum e_token_type
+{
+	WORD,
+	PIPE,
+	REDIR_IN,
+	REDIR_OUT,
+	REDIR_APPEND,
+	HEREDOC,
+	LOGICAL_AND,
+	LOGICAL_OR,
+	PAREN_OPEN,
+	PAREN_CLOSE,
+}	t_token_type;
+
+typedef enum e_quote
+{
+	NO_QUOTE,
+	SINGLE_QUOTE,
+	DOUBLE_QUOTE
+}	t_quote;
+
+typedef enum s_node_type
+{
+	NODE_COMMAND,
+	NODE_PIPE,
+	NODE_AND,
+	NODE_OR,
+	NODE_SUBSHELL
+}	t_node_type;
+
+typedef enum s_redir_type
+{
+	R_IN,
+	R_OUT,
+	R_APPEND,
+	R_HEREDOC
+}	t_redir_type;
+
+typedef struct s_redir
+{
+	t_redir_type	type;
+	char			*filename;
+	int				heredoc_fd;
+	struct s_redir	*next;
+}	t_redir;
+
+typedef struct s_command
+{
+	char			**args;
+	int				arg_count;
+	t_redir			*redirs;
+	int				heredoc_fd;
+	int				is_builtin;
+}	t_command;
+
+typedef struct s_token
+{
+	t_token_type	type;
+	char			*value;
+	struct s_token	*next;
+}	t_token;
+
+typedef struct s_token_list
+{
+	t_token	*head;
+	t_token	*tail;
+	int		size;
+}	t_token_list;
+
+typedef struct s_env
+{
+	char			*key;
+	char			*value;
+	struct s_env	*next;
+}	t_env;
+
+typedef struct s_ast
+{
+	t_node_type		type;
+	struct s_ast	*left;
+	struct s_ast	*right;
+	t_command		*cmd;
+}	t_ast;
+
+typedef struct s_args
+{
+	char			*arg;
+	struct s_args	*next;
+}	t_args;
+
+typedef struct s_expand
+{
+	t_args	**head;
+	t_quote	state;
+	t_env	*env;
+	char	*prefix;
+	char	*cur;
+	int		i;
+}	t_exp;
+
+typedef struct s_wild
+{
+	char	**parts;
+	int		have_start;
+	int		have_end;
+	int		full;
+}	t_wild;
+
+typedef struct s_shell
+{
+	t_env			*env;
+	t_token_list	*tokens;
+	t_ast			*ast;
+	int				last_status;
+}	t_shell;
+
+typedef struct s_quote_ctx
+{
+	char	*str;
+	char	*result;
+	int		i;
+	int		j;
+	t_quote	state;
+}	t_quote_ctx;
+
+extern volatile int			g_signal_status;
+
+#endif
